requirements.txt
# Python Version: 3.9.21
requests==2.32.3
python-dotenv==1.1.0